refactor(research): move CitedSourceSelection + research helpers to public surface
- Move CitedSourceSelection dataclass to notebooklm.types
- New notebooklm.research module exposes select_cited_sources, extract_report_urls, normalize_url
- _research.ResearchAPI.select_cited_sources/extract_report_urls preserved as delegating classmethods (backward compat)
- cli/helpers.py now imports from public surface
- Boundary plan: .sisyphus/plans/private-module-boundary.md (PR-A)
